# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Student:
        properties:
            active:
                type: boolean
                x-go-name: Active
            birthdate:
                description: The birthdate of the student
                pattern: DD/MM/YYYY
                type: string
                x-go-name: Birthdate
            cellphone:
                type: string
                x-go-name: Cellphone
            email:
                type: string
                x-go-name: Email
            gpa:
                format: float
                type: number
                x-go-name: GPA
            id:
                description: The id of the student
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            lastName:
                description: The last name of the student
                type: string
                x-go-name: LastName
            middleName:
                description: The middle name of the student
                type: string
                x-go-name: MiddleName
            name:
                description: The name of the student
                type: string
                x-go-name: Name
            totalCredits:
                format: int64
                type: integer
                x-go-name: TotalCredits
        required:
            - id
            - name
            - lastName
            - birthdate
        type: object
        x-go-package: github.com/Lemm8/CollegeManager/data
host: localhost
info:
    description: Documentation for Students API
    title: Students API
    version: 1.0.0
paths:
    /students:
        get:
            description: GetStudents returns the list of registred students from the data source (local list for now)
            operationId: listStudents
            responses:
                "200":
                    $ref: '#/responses/studentsResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Students
        post:
            description: Registers a student
            operationId: addStudent
            responses:
                "201":
                    $ref: '#/responses/created'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Students
    /students/{id}:
        delete:
            description: Delete a student
            operationId: deleteStudent
            parameters:
                - description: The id from the student
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Students
        get:
            description: Return student base on ID
            operationId: getStudent
            parameters:
                - description: The id from the student
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/studentResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Students
        put:
            description: Updates the information of a student
            operationId: updateStudent
            parameters:
                - description: The id from the student
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    $ref: '#/responses/noContent'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - Students
produces:
    - application/json
responses:
    created:
        description: Student Reigstered
    errorResponse:
        description: Generic error response
    noContent:
        description: No Content Response
    studentResponse:
        description: Single student
        schema:
            $ref: '#/definitions/Student'
    studentsResponse:
        description: List of students
        schema:
            items:
                $ref: '#/definitions/Student'
            type: array
schemes:
    - http
swagger: "2.0"